fix floating windows always raising even if we don't focus them, after changing tags
details:
Say we have 2 windows, on tag 0: one floating and one tiled. We focus on the tiled
window and then change tag (view another tag). When we come back to the
tag 0, the focused window will still be the tiled window, but the
floating window will be raised.
This wasn't intended but a leftover to the alwaysontop patch, so I also
applied bert patch always on top (called the same).