Monitor: remove sellt
similar modification as to the 'remove cringe toggle mechanism', that
one was for the tags and no functionality was removed.

This however, removes the ability to 'toggle' back a layout:
I won't ever see myself 'toggling' a layout (going back to the previous
one) since most of the time I cycle in between them.

adding a variable for oldlt and do a similar task as with seltags and
oldtags could probably be a 'better' solution, since it doesn't removes
stuff. But I found myself, agane, not using that functionality. might as
well get rid of it.

Let's C how it goes..

oh and all of this just cuz now it's easier to type c: less chars!1!1