manage: fix 'flicker' of swallowed windows
the arrange() call causes a call to the layout being used (window size
calculations and positions), which causes a 'flicker' effect for a
moment since there is a new mapped window.
'flicker': the new child window is view as the new master since it's a
new mapped window (XMapWindow()).
The swallow() function orders this in a way that the new child gets the
same position and dimention and the order in the stack list.

calling arrange() after swallow() is the correct order.