Pertag, make the '~0' ""tag"" (not really a tag) use the last index of
the array, which is LENGTH(tags).
Also add alltagslayout which is used for the '~0' tag. This tag is not
really a tag, it's all tags (TAGMASK) set to 1. Since Pertag is a
'hacky' but elegant solution to have 'pertag' (desktop like)
functionality, we can 'treat' this combination of all tags like a tag
("pertag tag").