getcmd: don't ignore, exec the command anyway
if the bar is hidden or showstatus is false, that would be handled in
run() anyway, by not poll()ing, and in drawbar(), by not drawing
anything (return;).

So by exec()ing the command, there would be data que to read() whenever
the bar is shown and statustext is true, which would cause to poll()
(see run()) and show the block with a text updated correctly.