manage: store the coordinates (hints) in sfx and sfy, if any
If none, the placeholder will be used to indicate that there isn't one
(since -9999 coordinates would not be realistic to reach with the mouse,
but possible with a keybinding).

This allows for windows that start in a custom size, which demwm forces
them to be tiled, restore it's original position (if any).

Example usecase: mpv will fit the window size arcording to the video
resolution, but this 'pixel perfect' window dimention will get lost and
unrecoverable thanks to the layout being forced on the window. With this
change pressing MOD+Shift+Space, togglefloating(), will make the window
floating and restore the position in which the window was before being
tiled.